% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudDemo-package.R
\docType{package}
\name{cloudDemo-package}
\alias{cloudDemo}
\alias{cloudDemo-package}
\title{cloudDemo: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/see24/cloudDemo}
  \item Report bugs at \url{https://github.com/see24/cloudDemo/issues}
}

}
\author{
\strong{Maintainer}: Sarah Endicott \email{sarah.endicott@canada.ca} (\href{https://orcid.org/0000-0001-9644-5343}{ORCID}) [copyright holder]

}
\keyword{internal}
