% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_mod.R
\name{do_mod}
\alias{do_mod}
\title{Run the model for any variable}
\usage{
do_mod(x)
}
\arguments{
\item{x}{variable name from mtcars}
}
\value{
data.frame with the variable name and r.squared value
}
\description{
Run a linear model of the variable vs mpg for the mtcars data set
}
