% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudDemo-package.R
\docType{package}
\name{cloudDemo-package}
\alias{cloudDemo}
\alias{cloudDemo-package}
\title{cloudDemo: Demo Cloud Project}
\description{
A repo for demonstrating how to use the cloud for a typical analysis in R
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LandSciTech/cloudDemo}
  \item \url{https://landscitech.github.io/cloudDemo/}
  \item Report bugs at \url{https://github.com/LandSciTech/cloudDemo/issues}
}

}
\author{
\strong{Maintainer}: Sarah Endicott \email{sarah.endicott@canada.ca} (\href{https://orcid.org/0000-0001-9644-5343}{ORCID}) [copyright holder]

}
\keyword{internal}
